device abstraction mostly done
