+++ src-tauri/Cargo.toml
@@ -66,6 +66,7 @@ md5 = "0.8.0"
 thiserror = "2.0.17"
 parking_lot = { version = "0.12.5", features = ["serde"] }
 directories = "6.0.0"
+enum_dispatch = "0.3.13"
 
 
 [target.'cfg(not(any(target_os = "android", target_os = "ios")))'.dependencies]
